campus-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
supabase
openai-agents
